requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Email integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
imaplib2==3.6
python-jose[cryptography]==3.3.0
exchangelib==5.1.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
tensorflow==2.14.0
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
nltk==3.8.1

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Security & Encryption
cryptography==41.0.7
bcrypt==4.1.1
passlib==1.7.4

# Web UI
jinja2==3.1.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
celery==5.3.4
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
httpx==0.25.2
pyyaml==6.0.1
rich==13.7.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0